server/requirements.txt
asgiref==3.9.1
Django==5.2.5
djangorestframework==3.16.1
djangorestframework-simplejwt==5.2.1
python-dotenv==1.1.1
setuptools<81
sqlparse==0.5.3
tzdata==2025.2
